% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_term_decomposition.R
\name{decompose_functional_term_splines}
\alias{decompose_functional_term_splines}
\title{Projection of the functional data in a finite dimensional B-spline basis}
\usage{
decompose_functional_term_splines(x, t, spl_n, spl_order)
}
\arguments{
\item{x}{Matrix of the functional covariates, where each row contains the
functions evaluated on a (common) grid.}

\item{t}{Numeric grid over which the functional covariates were evaluated.}

\item{spl_n}{basis size.}

\item{spl_order}{spline order (spl_order = 4 for cubic splines).}
}
\value{
A list with components:
\describe{
  \item{x_coeff}{Matrix with coefficients for projected functional observations,}
  \item{spl_basis_eval}{Matrix with spline basis evaluated on the same values as the functional observations.}
}
}
\description{
Given the size of a B-splines basis computes the corresponding coefficients
of the functional observations.
}
