% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/non_parametric_term_decomposition.R
\name{decompose_non_parametric_term_splines}
\alias{decompose_non_parametric_term_splines}
\title{Projection of the non-parametric term in a finite dimensional B-spline basis}
\usage{
decompose_non_parametric_term_splines(u, w, spl_n, spl_order)
}
\arguments{
\item{u}{Numeric vector with values of the explanatory variable that enters
the model non-parametrically.}

\item{w}{Numeric vector with varying coefficients for the non-parametric term
(w = 1 for non varying coefficients).}

\item{spl_n}{basis size.}

\item{spl_order}{spline order (spl_order = 4 for cubic splines).}
}
\value{
Matrix with spline basis evaluated on the same values as the
    functional observations.
}
\description{
Given the size of a B-splines basis computes the corresponding coefficients
of projection of the non-parametric term.
}
