% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goodness.R
\name{goodness}
\alias{goodness}
\title{Goodness of fit.}
\usage{
goodness(nn, scl, val, spl, freq, criterion)
}
\arguments{
\item{nn}{sample size.}

\item{scl}{scale estimate.}

\item{val}{minimization value.}

\item{spl}{number of splines1.}

\item{freq}{number of splines2.}

\item{criterion}{goodness of fit criterion.}
}
\value{
Goodness of fit for the selectec criterion.
}
\description{
Goodness of fit.
}
