% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbic.R
\name{rbic}
\alias{rbic}
\title{Robust BIC-type criterion}
\usage{
rbic(fit)
}
\arguments{
\item{n}{Sample size.}

\item{scl}{Scale estimate (numeric).}

\item{val}{Minimization value (numeric).}

\item{k_npt}{Number of parameters for the non-parametric term.}

\item{k_ft}{Number of parameters for the functional term.}
}
\value{
Information criterion for the selected criterion.
}
\description{
Computes a BIC-type information criterion for the current fit following the
proposal in Boente et al. (2020)
}
\references{
Boente, Graciela & Salibian-Barrera, Matias & Vena,
    Pablo. (2020). Robust estimation for semi-functional linear regression
    models. Computational Statistics & Data
    Analysis. 152. 107041. 10.1016/j.csda.2020.107041.
}
