% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FPLMBsplines.R
\name{FPLMBsplines}
\alias{FPLMBsplines}
\title{Best FPLMBsplines_fit given by a model selection criterion.}
\usage{
FPLMBsplines(
  y,
  x,
  u,
  t,
  range_freq = range_default,
  range_spl = range_default,
  norder = 4,
  fLoss = "lmrob",
  criterion = "bic1",
  trace = FALSE
)
}
\arguments{
\item{y}{the vector of scalar responses.}

\item{x}{a matrix of the functional covariates, where each row contains the
functions evaluated on a (common) grid.}

\item{u}{the values of the explanatory variable that enters the model
non-parametrically.}

\item{t}{the grid over which the functional covariates were evaluated.}

\item{range_freq}{a vector of B-spline basis sizes to try for the functional
regression coefficient.}

\item{range_spl}{a vector of B-spline basis sizes to try for the
non-parametric component.}

\item{norder}{the order of the B-Splines.}

\item{fLoss}{loss function to be minimized.}

\item{criterion}{criterion for model selection.}

\item{trace}{a logical argument indicating whether partial results are
printed.}
}
\value{
A list including the following components:
\itemize{
\item{fit}{fitted parameters}
\item{spl}{chosen number of splines for the non-parametric component}
\item{freq}{chosen number of splines for the funcitonal regression coefficient}
}
}
\description{
Fit a FPLM model for different spline basis sizes and picks the
    best one according to a specified model selection criterion.
}
\details{
Add some details regarding the criteria, valid ranges (asympotics)
    and floss options
}
\examples{

n <- 100
m <- 50
u <- runif(n)
t <- runif(m)
b <- function(x) x^3
g <- function(x) sin(x)
x <- matrix(rnorm(n * m), nrow = n)
y <- x \%*\% b(t) * min(diff(t)) + g(u) + rnorm(n, sd = 0.1)

FPLM_fit <- FPLMBsplines(y, x, u, t,
  range_freq = 4:13, range_spl = 4:13,
  norder = 4, fLoss = "ls", criterion = "bic1", trace = FALSE
)

par(mfrow = c(2, 1))
plot(t, FPLM_fit$fit$slope_fun, pch = 16)
plot(u, FPLM_fit$fit$eta_est, pch = 16)
}
