% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FPLMBsplines.R
\name{FPLMBsplines}
\alias{FPLMBsplines}
\title{Best FPLMBsplines fit by criterion.}
\usage{
FPLMBsplines(
  y,
  x,
  u,
  t,
  range_freq = range_default,
  range_spl = range_default,
  norder = 4,
  fLoss = "lmrob",
  criterion = "bic1",
  trace = FALSE
)
}
\arguments{
\item{y}{response}

\item{x}{description}

\item{u}{description}

\item{t}{description}

\item{range_freq}{description}

\item{range_spl}{description}

\item{norder}{Order of the splines.}

\item{fLoss}{Loss function}

\item{criterion}{Criterion for model selection}

\item{trace}{Logical.}
}
\value{
list(fit = fit_opt, spl = spl_opt, freq = freq_opt, u, t)
}
\description{
Best FPLMBsplines fit by criterion.
}
\examples{
Bla

}
