% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FPLMBsplines_fit.R
\name{FPLMBsplines_fit}
\alias{FPLMBsplines_fit}
\title{FPLM fit with splines}
\usage{
FPLMBsplines_fit(y, x, u, t, freq, spl, norder, fLoss)
}
\arguments{
\item{y}{vector of n responses}

\item{x}{matrix of functional covariates (n x p_1)}

\item{u}{vector of n observations for g(u_i)}

\item{t}{vector of "times" where X was observed "x[i, j] = X_i( t_j )"}

\item{freq}{description}

\item{spl}{description}

\item{norder}{description}

\item{fLoss}{loss function to be minimize (see Details)}
}
\value{
A list including several.
}
\description{
FPLM fit with splines
}
\details{
Valid options for the loss function are 'ls', 'lad', 'lmrob', 'huang'.
}
\examples{
Example.

}
\references{
Paper
}
