% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minimize.R
\name{minimize}
\alias{minimize}
\title{Minimization routine}
\usage{
minimize(
  y,
  x_coeff,
  u,
  w,
  k_npt,
  k_ft,
  fLoss,
  norder,
  rob.control = lmrob.control(trace.level = 0, nResample = 5000, tuning.psi = 3.443689,
    subsampling = "simple", rel.tol = 1e-05, refine.tol = 1e-05, k.max = 2000,
    maxit.scale = 2000, max.it = 2000)
)
}
\arguments{
\item{y}{Numeric vector of scalar responses.}

\item{x_coeff}{Numeric matrix with coefficients for the functional term
decomposition.}

\item{u}{Numeric vector with values of the explanatory variable that enters
the model non-parametrically.}

\item{w}{Numeric vector with varying coefficients for the non-parametric
term (w = 1 for non varying coefficients).}

\item{k_npt}{Number of splines for the non-parametric term.}

\item{k_ft}{Number of splines for the functional term.}

\item{fLoss}{String specifying the loss function. 'ls' for least squares,
'huang' for Huber, 'lmbrob' for MM-estimator.}

\item{norder}{Integer for spline order (norder = 4 for cubic splines).}

\item{rob.control}{Optional parameters if \code{fLoss == "lmbrob"}.}
}
\value{
A list with components:
\describe{
  \item{estimates_npt}{estimate for the non parametric term,}
  \item{estimates_ft}{estimate for the functional slope,}
  \item{value}{the minimum value of the loss,}
  \item{scale}{the scale estimate.}
}
}
\description{
This function is the minimization routine used internally by
\code{FPLMBsplines_fit()}, not meant for 'public' use.
}
\examples{

## Synthetic data
y <- rnorm(100)
x <- matrix(rnorm(5 * 100), ncol = 5)
u <- runif(100)
w <- 1

## Least-squares estimation
minimize(y, x, u, w, k_npt = 5, k_ft = 6, fLoss = "ls", norder = 4)

## M-estimation
minimize(y, x, u, w, k_npt = 5, k_ft = 6, fLoss = "huang", norder = 4)

## MM-estimation
minimize(y, x, u, w, k_npt = 5, k_ft = 6, fLoss = "lmrob", norder = 4)

## MM-estimation with 95\% of efficiency
minimize(y, x, u, w,
  k_npt = 5, k_ft = 6, fLoss = "lmrob", norder = 4,
  rob.control = list(tuning.psi = 4.685061)
)
}
