% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minimize.R
\name{minimize}
\alias{minimize}
\title{Minimization rutine}
\usage{
minimize(
  yy,
  xx_coef,
  uu,
  spl_kn,
  freq,
  fLoss,
  norder,
  rob.control = lmrob.control(trace.level = 0, nResample = 5000, tuning.psi = 3.443689,
    subsampling = "simple", rel.tol = 1e-05, refine.tol = 1e-05, k.max = 2000, maxit.scale
    = 2000, max.it = 2000)
)
}
\arguments{
\item{yy}{description}

\item{xx_coef}{description}

\item{uu}{description}

\item{spl_kn}{description}

\item{freq}{description}

\item{fLoss}{description}

\item{norder}{description}

\item{rob.control}{description}
}
\value{
Minimization stuff
}
\description{
This function is used internally by FPLMBsplines(), not meant
for 'public' use.
}
